✨ feat: Impl Repositories

+++ backend/timing/src/main/java/com/kkukku/timing/apis/challenge/repositories/ChallengeRepository.java
@@ -4,7 +4,7 @@ import com.kkukku.timing.apis.challenge.entities.ChallengeEntity;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChallengeRepository extends JpaRepository<ChallengeEntity, Integer> {
+public interface ChallengeRepository extends JpaRepository<ChallengeEntity, Long> {
 
     List<ChallengeEntity> findByMemberId(Integer memberId);
